backend/requirements/local.txt
-r base.txt
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.17.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
